# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Description of my API",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "List all articles",
                "description": "Returns a list of all articles",
                "operationId": "listArticles",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a specific article by ID",
                "description": "Returns a specific article based on ID",
                "operationId": "getArticleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}/like": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get the likes of an article",
                "description": "Get the likes of an article by ID",
                "operationId": "getArticleLikesById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article to retrieve likes for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}/comment": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get comments for an article",
                "description": "Returns the comments for a specific article by ID",
                "operationId": "e0a58a67415fc9b7b44c80a8e31ac8d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of article to get comments for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}/view": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get the views of a specific article",
                "description": "Returns the number of views for a specific article",
                "operationId": "views",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}